+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where Studio models with multiple assigned textures would only load the first texture.
+
 ## 0.4.8 - 2025-04-14
 ### Fixed
 * Fixed an issue where some Bricklink Studio files would fail to import due to incorrectly specified optional line commands.