+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## unreleased
+## 0.4.10 - 2025-07-11
 ### Fixed
 * Fixed an issue where the incorrect texture would be assigned to a face in some cases.